requirements.txt
fastapi
uvicorn
azure-cosmos
pydantic
pydantic[email]
email-validator
azure-cosmos
python-dotenv

